Move test ifname discovery utility to Utils
