+++ test/vintage_net/connectivity/internet_checker_test.exs
@@ -3,6 +3,7 @@ defmodule VintageNet.Connectivity.InternetCheckerTest do
   import ExUnit.CaptureLog
 
   alias VintageNet.Connectivity.InternetChecker
+  alias VintageNetTest.Utils
 
   test "rotate_list/1" do
     assert [1] == InternetChecker.rotate_list([1])
@@ -34,7 +35,7 @@ defmodule VintageNet.Connectivity.InternetCheckerTest do
       Application.start(:vintage_net)
     end)
 
-    ifname = get_ifname()
+    ifname = Utils.get_ifname_for_tests()
     property = ["interface", ifname, "connection"]
     lower_up = ["interface", ifname, "lower_up"]
 
@@ -75,27 +76,12 @@ defmodule VintageNet.Connectivity.InternetCheckerTest do
   test "deprecation warning when using old InternetConnectivityChecker" do
     messages =
       capture_log(fn ->
-        start_supervised!({VintageNet.Interface.InternetConnectivityChecker, get_ifname()})
+        start_supervised!(
+          {VintageNet.Interface.InternetConnectivityChecker, Utils.get_ifname_for_tests()}
+        )
       end)
 
     assert messages =~
              "VintageNet.Interface.InternetConnectivityChecker is now VintageNet.Connectivity.InternetChecker"
-  end
-
-  defp get_ifname() do
-    case :inet.getifaddrs() do
-      {:ok, addrs} ->
-        addrs
-        |> Enum.filter(&filter_interfaces/1)
-        |> List.first()
-        |> elem(0)
-        |> to_string()
-    end
-  end
-
-  defp filter_interfaces({[?l, ?o | _anything], _}), do: false
-
-  defp filter_interfaces({_ifname, fields}) do
-    Enum.member?(fields[:flags], :up) and fields[:addr] != nil
   end
 end
